Add test for vetting actor. Remove either and optional in order to simplify

+++ src/main/java/workshop/part1/Verdict.java
@@ -5,11 +5,11 @@ import workshop.fraudwordsservice.FraudWord;
 
 public class Verdict {
 
-    public final Long adId;
+    public final Integer adId;
     public final List<FraudWord> fraudWords;
     public final VerdictStatus verdictStatus;
 
-    public Verdict(Long adId, List<FraudWord> fraudWords, VerdictStatus verdictStatus) {
+    public Verdict(Integer adId, List<FraudWord> fraudWords, VerdictStatus verdictStatus) {
         this.adId = adId;
         this.fraudWords = fraudWords;
         this.verdictStatus = verdictStatus;
